+++ src/tooling/piral-cli-webpack/src/webpack/run-build-pilet.ts
@@ -33,7 +33,7 @@ async function run(
     externals,
     piral,
     version,
-    true,
+    false,
     sourceMaps,
     contentHash,
     minify,